requirements.txt
aiofiles==22.1.0
alembic==1.8.1
anyio==3.6.1
asgiref==3.6.0
asyncpg==0.27.0
attrs==22.1.0
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cleo==1.0.0a5
click==8.1.3
crashtest==0.3.1
cryptography==38.0.4
distlib==0.3.6
dnspython==2.2.1
dulwich==0.20.46
ecdsa==0.18.0
email-validator==1.3.0
exceptiongroup==1.0.4
Faker==15.3.4
fastapi==0.85.0
FastAPI-SQLAlchemy==0.2.1
filelock==3.8.0
h11==0.14.0
html5lib==1.1
idna==3.4
image==1.5.33
iniconfig==1.1.1
jaraco.classes==3.2.3
jsonschema==4.16.0
keyring==23.9.3
lockfile==0.12.2
Mako==1.2.3
MarkupSafe==2.1.1
more-itertools==8.14.0
msgpack==1.0.4
numpy==1.24.1
packaging==21.3
passlib==1.7.4
pexpect==4.8.0
Pillow==9.3.0
pkginfo==1.8.3
platformdirs==2.5.2
pluggy==1.0.0
poetry==1.2.1
poetry-core==1.2.0
poetry-plugin-export==1.1.1
psycopg2==2.9.4
ptyprocess==0.7.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
PyJWT==2.5.0
pylev==1.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.2.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.28.1
requests-toolbelt==0.9.1
rsa==4.9
shellingham==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlparse==0.4.3
starlette==0.20.4
tomli==2.0.1
tomlkit==0.11.5
typing_extensions==4.4.0
urllib3==1.26.12
utils==1.0.1
uvicorn==0.18.3
virtualenv==20.16.5
webencodings==0.5.1
xattr==0.9.9
